server/requirements.txt
Django==2.2.7
django-cors-headers==3.1.1
cloudinary==0.2.1
dj-database-url==0.5.0
django-heroku==0.3.1
djangorestframework==3.10.3
dlib==19.18.0
face-recognition==1.2.3
gunicorn==20.0.4
numpy==1.16.4
opencv-contrib-python==4.1.0.25
psycopg2==2.8.4
PyPDF2==1.26.0
reportlab==3.5.23
whitenoise==5.0.1